requirements.txt
requests
python-dotenv
descope
boto3
bcrypt
firebase-admin